Retry S3 uploads
We have been seeing occasional 500s from our save-cache operations. How
we're able to 500 the purported most reliable service on the internet is
beyond me, but the error indicates a retry may succeed, so we're
starting our investigation with that. If the operation continues to
consistently fail when retried, we can remove this and explore other
explanations (e.g. maybe it's our own bucket settings that have reduced
this reliability).

Control.Retry makes this very easy to implement. We're starting with its
default policy, which is a constant delay of 0.05s and a maximum of 5
attempts.